requirements.txt
botocore
boto3
python-dotenv
requests